@@ -1,4 +1,7 @@
-@preconcurrency import Foundation
+#if os(Linux)
+@preconcurrency import Dispatch
+#endif
+import Foundation
 import ConsoleKit
 import NIOConcurrencyHelpers
 
