@@ -1,4 +1,4 @@
-#if !canImport(Darwin)
+#if !canImport(Darwin) && compiler(<6.0)
 @preconcurrency import struct Foundation.URLComponents
 #else
 import struct Foundation.URLComponents
